fix(Network): 8 bugs across 5 modules — shutdown UAF, IRQL violation, race conditions
   NetworkFilter.c:
   - NF-C1 (CRITICAL): Move WFP unregister BEFORE sub-module shutdown to prevent
     in-flight classify callbacks from accessing freed sub-module memory
   - NF-H1 (HIGH): Add IRQL > APC_LEVEL guard at top of NfpProcessStreamData —
     stream callouts can fire at DISPATCH_LEVEL, push lock requires <= APC_LEVEL
   - NF-M1 (MEDIUM): Re-check DnsQueryCount inside exclusive DnsLock to close
     TOCTOU window that allowed exceeding NF_MAX_DNS_ENTRIES cap

   ConnectionTracker.c:
   - CT-C1 (HIGH): Replace non-atomic RefCount==1 + CtAddRef with
     InterlockedCompareExchange(&RefCount, 2, 1) in stale connection cleanup
   - CT-M1 (MEDIUM): On ObReferenceObjectByHandle failure, signal
     CleanupTerminate + wait for thread via handle before freeing tracker

   PortScanner.c:
   - PS-M1 (MEDIUM): Cast scan counters to LONGLONG before *100 to prevent
     32-bit integer overflow in percentage calculations

   ProtocolParser.c:
   - PP-H1 (HIGH): Replace InterlockedIncrement-then-check with CAS loop
     in PppTrackAllocation to prevent exceeding PP_MAX_CONCURRENT_ALLOCS

   DataExfiltration.c:
   - DX-H1 (HIGH): Route shutdown transfer cleanup through DxpDereferenceTransfer
     instead of direct ExFreeToNPagedLookasideList to maintain refcount contract

+++ PhantomSensor/PhantomSensor/Network/PortScanner.c
@@ -1894,30 +1894,31 @@ SspsDetermineScanType(
     }
 
     //
-    // Check stealth types first (most suspicious)
+    // Check stealth types first (most suspicious).
+    // Cast to LONGLONG before multiplication to prevent 32-bit overflow.
     //
     if (Snap->TcpNull > 0 &&
-        (Snap->TcpNull * 100 / Snap->TotalConnections) > 50) {
+        ((LONGLONG)Snap->TcpNull * 100 / Snap->TotalConnections) > 50) {
         return SsPsScan_TCPNULL;
     }
 
     if (Snap->TcpXmas > 0 &&
-        (Snap->TcpXmas * 100 / Snap->TotalConnections) > 50) {
+        ((LONGLONG)Snap->TcpXmas * 100 / Snap->TotalConnections) > 50) {
         return SsPsScan_TCPXMAS;
     }
 
     if (Snap->TcpFinOnly > 0 &&
-        (Snap->TcpFinOnly * 100 / Snap->TotalConnections) > 50) {
+        ((LONGLONG)Snap->TcpFinOnly * 100 / Snap->TotalConnections) > 50) {
         return SsPsScan_TCPFIN;
     }
 
     if (Snap->TcpSynOnly > 0 &&
-        (Snap->TcpSynOnly * 100 / Snap->TotalConnections) > 50) {
+        ((LONGLONG)Snap->TcpSynOnly * 100 / Snap->TotalConnections) > 50) {
         return SsPsScan_TCPSYN;
     }
 
     if (Snap->UdpConnections > 0 &&
-        (Snap->UdpConnections * 100 / Snap->TotalConnections) > 80) {
+        ((LONGLONG)Snap->UdpConnections * 100 / Snap->TotalConnections) > 80) {
         return SsPsScan_UDPScan;
     }
